build: enforce fatal warnings where CFLAGS cannot reset them (#109)
The tree selected -Wall -Werror for the RP2040 kernel and left -Wextra off, and several leaf Makefiles replace CFLAGS outright, so a root CFLAGS += -Werror would not have reached them. Severity now rides on CC through share/mk/warnings.mk for the ARM and MIPS target commands and through tools/Makefile.inc for the host tools, which a leaf can replace CFLAGS without escaping. HOST_CC stays a bare program name, because usr.bin/sort/Makefile passes it to a Python harness as argv[0].

Both RP2040 kernel configurations move to -Wall -Wextra -Werror, in the Makefile.rp2040 template and in the two tracked generated Makefiles, which regenerate byte-for-byte. The 231 instances that exposed -- 108 unused parameter, 100 signed comparison, 11 fallthrough, 8 setjmp clobber, 2 empty body, 2 enum range -- are repaired at the source: retained interface parameters are marked used, intentional fallthroughs annotated, and comparison types made consistent. Every cast is behavior preserving; uio_offset + uio_resid was already unsigned because uio_resid is u_int, and p_uid is short against a u_int uid_t either way, so the narrow-UID representation issue stands unfixed and unclaimed.

check-warning-policy-host and check-warning-policy-cross join the host and cross tiers. They evaluate each route's actual compiler command rather than grepping Makefiles for flag text, re-test the shared routes under CFLAGS=-O0, and count a negative probe only when the compile fails with the expected diagnostic. Calibrated both ways: the gate rejects a generated kernel Makefile with -Wextra removed and tools/Makefile.inc with -Werror removed, and passes the repaired tree.

Build, kernel, flash and every tier run clean with zero warnings on this branch merged with main, and the four CI jobs pass.

Shared userland makes its selected warnings fatal and is not globally -Wall -Wextra clean; that migration and the STM32/PIC32 matrix stay open. docs/research/warning-policy-audit.md records the audit and the boundaries.

+++ sys/kern/exec_elf.c
@@ -106,12 +106,12 @@ exec_elf_check(struct exec_params *epp)
      * Check that this is an ELF file that we can handle,
      * and do some sanity checks on the header
      */
-    if (epp->hdr_len < sizeof(struct elf_ehdr)) {
+    if (epp->hdr_len < (int)sizeof(struct elf_ehdr)) {
         DEBUG("\texec_elf_check(): error: wrong header length\n");
         DEBUG("\texec_elf_check(): end\n");
         return ENOEXEC;
     }
-    for (i = 0; i < sizeof elfident; i++) {
+    for (i = 0; i < (int)sizeof elfident; i++) {
         if (epp->hdr.elf.e_ident[i] !=  elfident[i]) {
             DEBUG("\texec_elf_check(): error: not an elf\n");
             DEBUG("\texec_elf_check(): end\n");
